refactor: applyNumberFormatter 메서드를 String Extension에 구현
